# webclient/swagger.yaml
openapi: 3.0.1
info:
  title: WebIO API
  version: v1
paths:
...
  /api/ChangeEvents/history:
    get:
      tags:
        - ChangeEvents
      operationId: ChangeEvents.GetHistory
      parameters:
        - name: start
          in: query
          schema:
            type: integer
            format: int32
            default: 0
        - name: count
          in: query
          schema:
            type: integer
            format: int32
            default: 100
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ChangeHistoryDtoQueryResultDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeHistoryDtoQueryResultDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ChangeHistoryDtoQueryResultDto'
      x-operation-name: getHistory
  /api/ChangeEvents:
    post:
      tags:
        - ChangeEvents
      operationId: ChangeEvents.Post
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PropertiesChangedSummaryDto'
          application/json:
            schema:
              $ref: '#/components/schemas/PropertiesChangedSummaryDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PropertiesChangedSummaryDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PropertiesChangedSummaryDto'
      responses:
        '200':
          description: Success
      x-operation-name: post
  /api/ChangeEvents/createDevice:
    post:
      tags:
        - ChangeEvents
      operationId: ChangeEvents.CreateDevice
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/DeviceAddedEventDto'
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAddedEventDto'
          text/json:
            schema:
              $ref: '#/components/schemas/DeviceAddedEventDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeviceAddedEventDto'
      responses:
        '200':
          description: Success
      x-operation-name: createDevice
  /api/ChangeEvents/deleteDevice:
    post:
      tags:
        - ChangeEvents
      operationId: ChangeEvents.DeleteDevice
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/DeviceDeletedDto'
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceDeletedDto'
          text/json:
            schema:
              $ref: '#/components/schemas/DeviceDeletedDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeviceDeletedDto'
      responses:
        '200':
          description: Success
      x-operation-name: deleteDevice
  /api/ChangeEvents/updateDevice:
    post:
      tags:
        - ChangeEvents
      operationId: ChangeEvents.UpdateDevice
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/DeviceUpdatedEventDto'
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdatedEventDto'
          text/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdatedEventDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DeviceUpdatedEventDto'
      responses:
        '200':
          description: Success
      x-operation-name: updateDevice
...
components:
  schemas:
    ChangeHistoryDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        details:
          nullable: true
      additionalProperties: false
    ChangeHistoryDtoQueryResultDto:
      type: object
      properties:
        state:
          $ref: '#/components/schemas/QueryResultState'
        start:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChangeHistoryDto'
          nullable: true
      additionalProperties: false
...
    DeviceAddedEventDto:
      type: object
      properties:
        comment:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        deviceType:
          type: string
          nullable: true
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/InterfaceTemplateSelectionDto'
          nullable: true
        useSt2022_7:
          type: boolean
      additionalProperties: false
    DeviceDeletedDto:
      type: object
      properties:
        comment:
          type: string
          nullable: true
        deviceId:
          type: string
          format: uuid
      additionalProperties: false
...
    DeviceUpdatedEventDto:
      type: object
      properties:
        comment:
          type: string
          nullable: true
        deviceId:
          type: string
          format: uuid
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/InterfaceTemplateSelectionDto'
          nullable: true
      additionalProperties: false
...
    InterfaceTemplateSelectionDto:
      type: object
      properties:
        templateName:
          type: string
          nullable: true
        audioSend:
          type: integer
          format: int32
        audioReceive:
          type: integer
          format: int32
        videoSend:
          type: integer
          format: int32
        videoReceive:
          type: integer
          format: int32
        ancillarySend:
          type: integer
          format: int32
        ancillaryReceive:
          type: integer
          format: int32
      additionalProperties: false
    PropertiesChangedSummaryDto:
      type: object
      properties:
        comment:
          type: string
          nullable: true
        changedEvents:
          type: array
          items:
            $ref: '#/components/schemas/PropertyChangedEventDto'
          nullable: true
      additionalProperties: false
    PropertyChangedEventDto:
      type: object
      properties:
        property:
          type: string
          nullable: true
        newValue:
          type: string
          nullable: true
        entity:
          type: string
          nullable: true
        device:
          type: string
          nullable: true
        entityType:
          type: string
          nullable: true
      additionalProperties: false
    QueryResultState:
      enum:
        - Success
      type: string
...